requirements.txt
# Core dependencies for Elliptic GNN Baselines
# Python >= 3.10 required

# Deep Learning
torch>=2.0.0
torch-geometric>=2.3.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Gradient Boosting
xgboost>=2.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
pyyaml>=6.0
tqdm>=4.65.0

# Jupyter
notebook>=7.0.0
ipykernel>=6.25.0

# Testing (optional)
pytest>=7.4.0
